pyproject.toml
[tool.poetry]
name = "acr-hiring-task"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.0"
openai = "1.11.1"
python-dotenv = "1.0.1"
pymupdf = "1.23.21"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.1.1"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
target-version = ['py311']
